remove provider and use cubits in theme , fontsize and rectire seperate the settings screen to small widgets handle pubspec remove unused fonts and images
