+++ lib/features/settings/view/settings_view.dart
@@ -1,0 +1,16 @@
+import 'package:flutter/material.dart';
+
+import 'widgets/settings_content.dart';
+
+class SettingsView extends StatelessWidget {
+  const SettingsView({super.key});
+
+  @override
+  Widget build(BuildContext context) => Directionality(
+    textDirection: TextDirection.rtl,
+    child: Scaffold(
+      appBar: AppBar(title: const Text('الإعدادات')),
+      body: const SettingsContent(),
+    ),
+  );
+}
