+++ lib/features/settings/view/widgets/section_card.dart
@@ -1,0 +1,26 @@
+import 'package:flutter/material.dart';
+
+class SectionCard extends StatelessWidget {
+  const SectionCard({required this.title, required this.child, super.key});
+  final String title;
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    final theme = Theme.of(context);
+
+    return Card(
+      child: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Text(title, style: theme.textTheme.titleLarge),
+            const SizedBox(height: 16),
+            child,
+          ],
+        ),
+      ),
+    );
+  }
+}
