+++ lib/main.dart
@@ -1,23 +1,23 @@
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
+import 'features/settings/view_model/font_size/font_size_cubit.dart';
+import 'features/settings/view_model/rectire/rectire_cubit.dart';
+import 'features/settings/view_model/theme/theme_cubit.dart';
 import 'core/main/main_content/app_content.dart';
 import 'core/main/main_content/app_initializer.dart';
 import 'core/main/main_content/error_app.dart';
-import 'screens/settings/provider/rectire_provider.dart';
-import 'screens/settings/provider/theme_provider.dart';
-import 'screens/settings/provider/font_size_provider.dart';
 
 class MuslimApp extends StatelessWidget {
   const MuslimApp({super.key});
 
   @override
-  Widget build(BuildContext context) => MultiProvider(
+  Widget build(BuildContext context) => MultiBlocProvider(
     providers: [
-      ChangeNotifierProvider(create: (_) => FontSizeProvider()),
-      ChangeNotifierProvider(create: (_) => ThemeProvider()),
-      ChangeNotifierProvider(create: (_) => ReciterProvider()),
+      BlocProvider(create: (_) => FontSizeCubit()),
+      BlocProvider(create: (_) => ThemeCubit()),
+      BlocProvider(create: (_) => ReciterCubit()),
     ],
     child: const AppContent(),
   );
