+++ pubspec.yaml
@@ -36,6 +36,4 @@ flutter:
   uses-material-design: true
 
   assets:
-    - assets/images/compass.png
-    - assets/images/arrow.png
     - assets/images/kaaba.png
